Bug 1858378 - Fix flaky review quality check navigation middleware unit test
